fix : list -> studyList

+++ src/main/java/com/baeker/baeker/studyRule/StudyRuleController.java
@@ -37,7 +37,6 @@ public class StudyRuleController {
     @PreAuthorize("isAuthenticated()")
     public String showForm(Model model, @PathVariable Long id, StudyRuleForm studyRuleForm) {
         List<Rule> ruleList = ruleService.getRuleList();
-        model.addAttribute("studyId", id);
         model.addAttribute("ruleList", ruleList);
         return "studyRule/create";
     }
@@ -47,7 +46,7 @@ public class StudyRuleController {
     public String create(@PathVariable("id") Long id,@RequestParam("rule") Rule rule,
                          @Valid StudyRuleForm studyRuleForm, BindingResult bindingResult) {
 
-        RsData<Study> exist = studyRuleService.isExist(rq, id);
+        RsData<Study> exist = studyRuleService.verificationLeader(rq, id);
         if (exist.isFail()) {
             return rq.historyBack("스터디 리더가 아닙니다.");
         }
@@ -57,7 +56,7 @@ public class StudyRuleController {
         if (rsData.isFail() || bindingResult.hasErrors()) {
             return rq.historyBack(rsData);
         }
-        return rq.redirectWithMsg("/studyRule/list", rsData.getMsg());
+        return rq.redirectWithMsg(String.format("/studyRule/list/%s", id), rsData.getMsg());
     }
 
     /**
@@ -67,6 +66,13 @@ public class StudyRuleController {
     @GetMapping("/modify/{id}")
     @PreAuthorize("isAuthenticated()")
     public String showModify(Model model,@PathVariable("id") Long id, StudyRuleForm studyRuleForm) {
+        Long studyId = studyRuleService.getStudyId(id);
+
+        RsData<Study> exist = studyRuleService.verificationLeader(rq, studyId);
+        if (exist.isFail()) {
+            return rq.historyBack("스터디 리더가 아닙니다.");
+        }
+
         List<Rule> ruleList = ruleService.getRuleList();
         RsData<StudyRule> rsData = studyRuleService.getStudyRule(id);
         if (rsData.isSuccess()) {
@@ -82,6 +88,13 @@ public class StudyRuleController {
     @PostMapping("/modify/{id}")
     @PreAuthorize("isAuthenticated()")
     public String modify(@PathVariable("id") Long id, @Valid StudyRuleForm studyRuleForm, BindingResult bindingResult) {
+        Long studyId = studyRuleService.getStudyId(id);
+
+        RsData<Study> exist = studyRuleService.verificationLeader(rq, studyId);
+        if (exist.isFail()) {
+            return rq.historyBack("스터디 리더가 아닙니다.");
+        }
+
         RsData<StudyRule> rsData = studyRuleService.getStudyRule(id);
         if (rsData.isFail() || bindingResult.hasErrors()) {
             return rq.historyBack(rsData);
@@ -96,6 +109,13 @@ public class StudyRuleController {
     @DeleteMapping("/delete/{id}")
     @PreAuthorize("isAuthenticated()")
     public String delete(@PathVariable("id") Long id) {
+        Long studyId = studyRuleService.getStudyId(id);
+
+        RsData<Study> exist = studyRuleService.verificationLeader(rq, studyId);
+        if (exist.isFail()) {
+            return rq.historyBack("스터디 리더가 아닙니다.");
+        }
+
         RsData<StudyRule> rsData = studyRuleService.getStudyRule(id);
         if (studyRuleService.delete(rsData.getData(), rsData.getData().getStudy().getLeader(), rq.getMember().getNickName()).isSuccess()) {
             return rq.redirectWithMsg("/studyRule/list", "삭제 되었습니다.");
@@ -106,10 +126,10 @@ public class StudyRuleController {
     /**
      * 조회
      */
-    @GetMapping("/list")
+    @GetMapping("/list/{id}")
     @PreAuthorize("isAuthenticated()")
-    public String showList(Model model, @RequestParam(defaultValue = "0") int page){
-        Page<StudyRule> paging = studyRuleService.getList(page);
+    public String showList(Model model, @PathVariable("id") Long id, @RequestParam(defaultValue = "0") int page){
+        Page<StudyRule> paging = studyRuleService.getList(page, id);
         model.addAttribute("paging", paging);
         return "studyRule/studyRuleList";
     }
